refactored package names to have src as source folder
