+++ src/api/ApiHandlerClient.java
@@ -1,4 +1,4 @@
-package src.api;
+package api;
 
 import se.michaelthelin.spotify.SpotifyApi;
 import se.michaelthelin.spotify.model_objects.credentials.ClientCredentials;
