+++ src/entities/Leaderboard.java
@@ -1,4 +1,4 @@
-package src.entities;
+package entities;
 public class Leaderboard {
     private Player[] players;
