+++ src/entities/Player.java
@@ -1,4 +1,4 @@
-package src.entities;
+package entities;
 public class Player {
 
     private String name;
