+++ src/entities/Question.java
@@ -1,4 +1,4 @@
-package src.entities;
+package entities;
 
 public class Question {
     public String question;
