--- src/interface_adapter/QuizViewModel.java
@@ -1,4 +1,0 @@
-package src.interface_adapter;
-
-public class QuizViewModel {
-}
