--- src/interface_adapter/take_quiz/TakeQuizController.java
@@ -1,4 +1,0 @@
-package src.interface_adapter.take_quiz;
-
-public class TakeQuizController {
-}
