+++ src/use_case/take_quiz/TakeQuizOutputBoundary.java
@@ -1,4 +1,4 @@
-package src.use_case.take_quiz;
+package use_case.take_quiz;
 
 public interface TakeQuizOutputBoundary {
     void prepareSuccessView(TakeQuizOutputData takeQuizOutputData);
