--- src/view/TakeQuizView.java
@@ -1,5 +1,0 @@
-package src.view;
-
-public class TakeQuizView {
-
-}
